+++ python-sdk/example_dags/example_datasets.py
@@ -30,11 +30,11 @@ input_file = File(path="https://raw.githubusercontent.com/astronomer/astro-sdk/m
 # [END dataset_file]
 
 # [START dataset_table]
-imdb_movies_table = Table(name="imdb_movies", conn_id="sqlite_default")
+imdb_movies_table = Table(name="imdb_movies", conn_id="postgres_conn")
 # [END dataset_table]
 
-top_animations_table = Table(name="top_animation", conn_id="sqlite_default")
-START_DATE = datetime(2022, 9, 1)
+top_animations_table = Table(name="top_animation", conn_id="postgres_conn")
+START_DATE = datetime(2016, 9, 1)
 
 
 @aql.transform()