+++ python-sdk/src/astro/databases/snowflake.py
@@ -66,6 +66,8 @@ NATIVE_LOAD_SUPPORTED_FILE_LOCATIONS = (FileLocation.GS, FileLocation.S3)
 NATIVE_AUTODETECT_SCHEMA_SUPPORTED_FILE_TYPES = {FileType.PARQUET}
 NATIVE_AUTODETECT_SCHEMA_SUPPORTED_FILE_LOCATIONS = {FileLocation.GS, FileLocation.S3}
 
+COPY_INTO_COMMAND_FAIL_STATUS = "LOAD_FAILED"
+
 
 @dataclass
 class SnowflakeFileFormat:
@@ -250,13 +252,21 @@ class SnowflakeDatabase(BaseDatabase):
     )
     DEFAULT_SCHEMA = SNOWFLAKE_SCHEMA
 
-    def __init__(self, conn_id: str = DEFAULT_CONN_ID):
+    def __init__(self, conn_id: str = DEFAULT_CONN_ID, table: BaseTable | None = None):
         super().__init__(conn_id)
+        self.table = table
 
     @property
     def hook(self) -> SnowflakeHook:
         """Retrieve Airflow hook to interface with the snowflake database."""
-        return SnowflakeHook(snowflake_conn_id=self.conn_id)
+        kwargs = {}
+        _hook = SnowflakeHook(snowflake_conn_id=self.conn_id)
+        if self.table and self.table.metadata:
+            if _hook.database is None and self.table.metadata.database:
+                kwargs.update({"database": self.table.metadata.database})
+            if _hook.schema is None and self.table.metadata.schema:
+                kwargs.update({"schema": self.table.metadata.schema})
+        return SnowflakeHook(snowflake_conn_id=self.conn_id, **kwargs)
 
     @property
     def sql_type(self) -> str:
@@ -508,7 +518,7 @@ class SnowflakeDatabase(BaseDatabase):
         file: File | None = None,
         dataframe: pd.DataFrame | None = None,
         columns_names_capitalization: ColumnCapitalization = "original",
-    ) -> None:
+    ) -> None:  # skipcq PYL-W0613
         """
         Create a SQL table, automatically inferring the schema using the given file.
         Overriding default behaviour and not using the `prep_table` since it doesn't allow the adding quotes.
@@ -542,7 +552,9 @@ class SnowflakeDatabase(BaseDatabase):
         # Since this method is used by both native and pandas path we cannot skip this step.
         self.truncate_table(table)
 
-    def is_native_load_file_available(self, source_file: File, target_table: BaseTable) -> bool:
+    def is_native_load_file_available(
+        self, source_file: File, target_table: BaseTable  # skipcq PYL-W0613, PYL-R0201
+    ) -> bool:
         """
         Check if there is an optimised path for source to destination.
 
@@ -562,7 +574,7 @@ class SnowflakeDatabase(BaseDatabase):
         if_exists: LoadExistStrategy = "replace",
         native_support_kwargs: dict | None = None,
         **kwargs,
-    ):
+    ):  # skipcq PYL-W0613
         """
         Load the content of a file to an existing Snowflake table natively by:
         - Creating a Snowflake external stage
@@ -594,11 +606,28 @@ class SnowflakeDatabase(BaseDatabase):
         table_name = self.get_table_qualified_name(target_table)
         file_path = os.path.basename(source_file.path) or ""
         sql_statement = f"COPY INTO {table_name} FROM @{stage.qualified_name}/{file_path}"
+
+        # Below code is added due to breaking change in apache-airflow-providers-snowflake==3.2.0,
+        # we need to pass handler param to get the rows. But in version apache-airflow-providers-snowflake==3.1.0
+        # if we pass the handler provider raises an exception AttributeError
         try:
-            self.hook.run(sql_statement)
-        except (ValueError, AttributeError) as exe:
+            rows = self.hook.run(sql_statement, handler=lambda cur: cur.fetchall())
+        except AttributeError:
+            try:
+                rows = self.hook.run(sql_statement)
+            except (AttributeError, ValueError) as exe:
+                raise DatabaseCustomError from exe
+        except ValueError as exe:
             raise DatabaseCustomError from exe
+
+        self.evaluate_results(rows)
         self.drop_stage(stage)
+
+    @staticmethod
+    def evaluate_results(rows):
+        """check the error state returned by snowflake when running `copy into` query."""
+        if any(row["status"] == COPY_INTO_COMMAND_FAIL_STATUS for row in rows):
+            raise DatabaseCustomError(rows)
 
     def load_pandas_dataframe_to_table(
         self,
@@ -636,7 +665,7 @@ class SnowflakeDatabase(BaseDatabase):
 
     def get_sqlalchemy_template_table_identifier_and_parameter(
         self, table: BaseTable, jinja_table_identifier: str
-    ) -> tuple[str, str]:
+    ) -> tuple[str, str]:  # skipcq PYL-R0201
         """
         During the conversion from a Jinja-templated SQL query to a SQLAlchemy query, there is the need to
         convert a Jinja table identifier to a safe SQLAlchemy-compatible table identifier.
@@ -883,7 +912,7 @@ class SnowflakeDatabase(BaseDatabase):
             identifier_enclosure = '"'
 
         constraints = ",".join([f"{identifier_enclosure}{p}{identifier_enclosure}" for p in parameters])
-        sql = "ALTER TABLE {{table}} ADD CONSTRAINT airflow UNIQUE (%s)" % constraints
+        sql = "ALTER TABLE {{table}} ADD CONSTRAINT airflow UNIQUE (%s)" % constraints  # skipcq PYL-C0209
         return sql
 
     def openlineage_dataset_name(self, table: BaseTable) -> str: