+++ python-sdk/src/astro/files/types/csv.py
@@ -12,7 +12,10 @@ from astro.utils.dataframe import convert_columns_names_capitalization
 class CSVFileType(FileType):
     """Concrete implementation to handle CSV file type"""
 
-    def export_to_dataframe(self, stream, columns_names_capitalization="original", **kwargs) -> pd.DataFrame:
+    # We need skipcq because it's a method overloading so we don't want to make it a static method
+    def export_to_dataframe(
+        self, stream, columns_names_capitalization="original", **kwargs
+    ) -> pd.DataFrame:  # skipcq PYL-R0201
         """read csv file from one of the supported locations and return dataframe
 
         :param stream: file stream object
@@ -25,7 +28,8 @@ class CSVFileType(FileType):
         )
         return df
 
-    def create_from_dataframe(self, df: pd.DataFrame, stream: io.TextIOWrapper) -> None:
+    # We need skipcq because it's a method overloading so we don't want to make it a static method
+    def create_from_dataframe(self, df: pd.DataFrame, stream: io.TextIOWrapper) -> None:  # skipcq PYL-R0201
         """Write csv file to one of the supported locations
 
         :param df: pandas dataframe