+++ python-sdk/src/astro/sql/operators/export_file.py
@@ -5,11 +5,22 @@ from typing import Any
 import pandas as pd
 from airflow.decorators.base import get_unique_task_id
 from airflow.models.xcom_arg import XComArg
+from openlineage.client.facet import (
+    BaseFacet,
+    DataQualityMetricsInputDatasetFacet,
+    DataSourceDatasetFacet,
+    OutputStatisticsOutputDatasetFacet,
+    SchemaDatasetFacet,
+    SchemaField,
+)
+from openlineage.client.run import Dataset as OpenlineageDataset
 
 from astro.airflow.datasets import kwargs_with_datasets
 from astro.constants import ExportExistsStrategy
 from astro.databases import create_database
 from astro.files import File
+from astro.lineage.extractor import OpenLineageFacets
+from astro.lineage.facets import ExportFileFacet
 from astro.sql.operators.base_operator import AstroSQLBaseOperator
 from astro.table import BaseTable, Table
 from astro.utils.typing_compat import Context
@@ -41,14 +52,14 @@ class ExportFileOperator(AstroSQLBaseOperator):
             datasets["input_datasets"] = input_data
         super().__init__(**kwargs_with_datasets(kwargs=kwargs, **datasets))
 
-    def execute(self, context: Context) -> File:
+    def execute(self, context: Context) -> File:  # skipcq PYL-W0613
         """Write SQL table to csv/parquet on local/S3/GCS.
 
         Infers SQL database type based on connection.
         """
         # Infer db type from `input_conn_id`.
         if isinstance(self.input_data, BaseTable):
-            database = create_database(self.input_data.conn_id)
+            database = create_database(self.input_data.conn_id, table=self.input_data)
             self.input_data = database.populate_table_metadata(self.input_data)
             df = database.export_table_to_pandas_dataframe(self.input_data)
         elif isinstance(self.input_data, pd.DataFrame):
@@ -61,6 +72,71 @@ class ExportFileOperator(AstroSQLBaseOperator):
             return self.output_file
         else:
             raise FileExistsError(f"{self.output_file.path} file already exists.")
+
+    def get_openlineage_facets(self, task_instance) -> OpenLineageFacets:  # skipcq: PYL-W0613
+        """
+        Collect the input, output, job and run facets for export file operator
+        """
+        input_dataset: list[OpenlineageDataset] = []
+        if isinstance(self.input_data, BaseTable) and self.input_data.openlineage_emit_temp_table_event():
+            input_uri = (
+                f"{self.input_data.openlineage_dataset_namespace()}"
+                f"://{self.input_data.openlineage_dataset_name()}"
+            )
+            input_dataset = [
+                OpenlineageDataset(
+                    namespace=self.input_data.openlineage_dataset_namespace(),
+                    name=self.input_data.openlineage_dataset_name(),
+                    facets={
+                        "dataSource": DataSourceDatasetFacet(
+                            name=self.input_data.openlineage_dataset_name(), uri=input_uri
+                        ),
+                        "schema": SchemaDatasetFacet(
+                            fields=[
+                                SchemaField(
+                                    name=self.input_data.metadata.schema,
+                                    type=self.input_data.metadata.database,
+                                )
+                            ]
+                        ),
+                        "dataQualityMetrics": DataQualityMetricsInputDatasetFacet(
+                            rowCount=self.input_data.row_count, columnMetrics={}
+                        ),
+                    },
+                )
+            ]
+        output_uri = (
+            f"{self.output_file.openlineage_dataset_namespace}"
+            f"://{self.output_file.openlineage_dataset_name}"
+        )
+        output_dataset = [
+            OpenlineageDataset(
+                namespace=self.output_file.openlineage_dataset_namespace,
+                name=self.output_file.openlineage_dataset_name,
+                facets={
+                    "output_file_facet": ExportFileFacet(
+                        filepath=self.output_file.path,
+                        file_size=self.output_file.size,
+                        file_type=self.output_file.type.name,
+                        if_exists=self.if_exists,
+                    ),
+                    "dataSource": DataSourceDatasetFacet(name=self.output_file.path, uri=output_uri),
+                    "outputStatistics": OutputStatisticsOutputDatasetFacet(
+                        rowCount=self.input_data.row_count
+                        if isinstance(self.input_data, BaseTable)
+                        else len(self.input_data),
+                        size=self.output_file.size,
+                    ),
+                },
+            )
+        ]
+
+        run_facets: dict[str, BaseFacet] = {}
+        job_facets: dict[str, BaseFacet] = {}
+
+        return OpenLineageFacets(
+            inputs=input_dataset, outputs=output_dataset, run_facets=run_facets, job_facets=job_facets
+        )
 
 
 def export_file(